+++ textfile/widget.go
@@ -2,14 +2,18 @@ package textfile
 
 import (
 	"bytes"
+	"fmt"
 	"io/ioutil"
+	"log"
 	"os"
 	"path/filepath"
+	"time"
 
 	"github.com/alecthomas/chroma/formatters"
 	"github.com/alecthomas/chroma/lexers"
 	"github.com/alecthomas/chroma/styles"
 	"github.com/gdamore/tcell"
+	"github.com/radovskyb/watcher"
 	"github.com/rivo/tview"
 	"github.com/senorprogrammer/wtf/wtf"
 )
@@ -18,23 +22,32 @@ const HelpText = `
   Keyboard commands for Textfile:
 
     /: Show/hide this help window
+    h: Previous text file
+    l: Next text file
     o: Open the text file in the operating system
+
+    arrow left:  Previous text file
+    arrow right: Next text file
 `
 
 type Widget struct {
 	wtf.HelpfulWidget
+	wtf.MultiSourceWidget
 	wtf.TextWidget
-
-	filePath string
 }
 
 func NewWidget(app *tview.Application, pages *tview.Pages) *Widget {
 	widget := Widget{
-		HelpfulWidget: wtf.NewHelpfulWidget(app, pages, HelpText),
-		TextWidget:    wtf.NewTextWidget("TextFile", "textfile", true),
-
-		filePath: wtf.Config.UString("wtf.mods.textfile.filePath"),
+		HelpfulWidget:     wtf.NewHelpfulWidget(app, pages, HelpText),
+		MultiSourceWidget: wtf.NewMultiSourceWidget("textfile", "filePath", "filePaths"),
+		TextWidget:        wtf.NewTextWidget("TextFile", "textfile", true),
 	}
+
+	// Don't use a timer for this widget, watch for filesystem changes instead
+	widget.RefreshInt = 0
+
+	widget.LoadSources()
+	widget.SetDisplayFunction(widget.display)
 
 	widget.HelpfulWidget.SetView(widget.View)
 
@@ -42,33 +55,42 @@ func NewWidget(app *tview.Application, pages *tview.Pages) *Widget {
 	widget.View.SetWordWrap(true)
 	widget.View.SetInputCapture(widget.keyboardIntercept)
 
+	go widget.watchForFileChanges()
+
 	return &widget
 }
 
 /* -------------------- Exported Functions -------------------- */
 
+// Refresh is only called once on start-up. Its job is to display the
+// text files that first time. After that, the watcher takes over
 func (widget *Widget) Refresh() {
-	widget.UpdateRefreshedAt()
-	widget.View.SetTitle(widget.ContextualTitle(widget.fileName()))
+	widget.display()
+}
 
-	var text string
+/* -------------------- Unexported Functions -------------------- */
+
+func (widget *Widget) display() {
+	title := fmt.Sprintf("[green]%s[white]", widget.CurrentSource())
+	widget.View.SetTitle(widget.ContextualTitle(title))
+
+	text := wtf.SigilStr(len(widget.Sources), widget.Idx, widget.View) + "\n"
+
 	if wtf.Config.UBool("wtf.mods.textfile.format", false) {
-		text = widget.formattedText()
+		text = text + widget.formattedText()
 	} else {
-		text = widget.plainText()
+		text = text + widget.plainText()
 	}
 
 	widget.View.SetText(text)
 }
 
-/* -------------------- Unexported Functions -------------------- */
-
 func (widget *Widget) fileName() string {
-	return filepath.Base(widget.filePath)
+	return filepath.Base(widget.CurrentSource())
 }
 
 func (widget *Widget) formattedText() string {
-	filePath, _ := wtf.ExpandHomeDir(widget.filePath)
+	filePath, _ := wtf.ExpandHomeDir(widget.CurrentSource())
 
 	file, err := os.Open(filePath)
 	if err != nil {
@@ -99,9 +121,11 @@ func (widget *Widget) formattedText() string {
 }
 
 func (widget *Widget) plainText() string {
-	filePath, _ := wtf.ExpandHomeDir(widget.filePath)
+	filePath, _ := wtf.ExpandHomeDir(widget.CurrentSource())
 
-	text, err := ioutil.ReadFile(filePath) // just pass the file name
+	fmt.Println(filePath)
+
+	text, err := ioutil.ReadFile(filePath)
 	if err != nil {
 		return err.Error()
 	}
@@ -113,10 +137,60 @@ func (widget *Widget) keyboardIntercept(event *tcell.EventKey) *tcell.EventKey {
 	case "/":
 		widget.ShowHelp()
 		return nil
+	case "h":
+		widget.Prev()
+		return nil
+	case "l":
+		widget.Next()
+		return nil
 	case "o":
-		wtf.OpenFile(widget.filePath)
+		wtf.OpenFile(widget.CurrentSource())
 		return nil
 	}
 
+	switch event.Key() {
+	case tcell.KeyLeft:
+		widget.Prev()
+		return nil
+	case tcell.KeyRight:
+		widget.Next()
+		return nil
+	default:
+		return event
+	}
+
 	return event
+}
+
+func (widget *Widget) watchForFileChanges() {
+	watch := watcher.New()
+	watch.FilterOps(watcher.Write)
+
+	go func() {
+		for {
+			select {
+			case <-watch.Event:
+				widget.display()
+			case err := <-watch.Error:
+				log.Fatalln(err)
+			case <-watch.Closed:
+				return
+			}
+		}
+	}()
+
+	// Watch each textfile for changes
+	for _, source := range widget.Sources {
+		fullPath, err := wtf.ExpandHomeDir(source)
+		if err == nil {
+			if err := watch.Add(fullPath); err != nil {
+				log.Fatalln(err)
+			}
+		}
+	}
+
+	// Start the watching process - it'll check for changes every 100ms.
+	if err := watch.Start(time.Millisecond * 100); err != nil {
+		log.Fatalln(err)
+	}
 }