+++ wtf/focus_tracker.go
@@ -25,20 +25,25 @@ type FocusTracker struct {
 // AssignHotKeys assigns an alphabetic keyboard character to each focusable
 // widget so that the widget can be brought into focus by pressing that keyboard key
 func (tracker *FocusTracker) AssignHotKeys() {
-	if !tracker.withShortcuts() {
+	if !tracker.useNavShortcuts() {
 		return
 	}
 
-	i := 0
+	i := 1
 
 	for _, focusable := range tracker.focusables() {
-		focusable.SetFocusChar(string('a' + i))
+		// Don't have nav characters > "9"
+		if i >= 10 {
+			break
+		}
+
+		focusable.SetFocusChar(string('0' + i))
 		i++
 	}
 }
 
 func (tracker *FocusTracker) FocusOn(char string) bool {
-	if !tracker.withShortcuts() {
+	if !tracker.useNavShortcuts() {
 		return false
 	}
 
@@ -176,6 +181,6 @@ func (tracker *FocusTracker) increment() {
 	}
 }
 
-func (tracker *FocusTracker) withShortcuts() bool {
+func (tracker *FocusTracker) useNavShortcuts() bool {
 	return Config.UBool("wtf.navigation.shortcuts", true)
 }